+++ R/utils.R
@@ -208,4 +208,10 @@ dyn_register_s3_method <- function(
   )
 }
 
+
+
+last <- function(x){
+  x[length(x)]
+}
+
 # nocov end